Extract expression classes to own files
