+++ src/controllers/filters/parser/Types.hpp
@@ -1,9 +1,7 @@
 #pragma once
 
-#include <QRegularExpression>
-#include <set>
-
 #include <memory>
+#include <set>
 
 namespace chatterino {
 
@@ -12,6 +10,8 @@ struct Message;
 }
 
 namespace filterparser {
+
+class Expression;
 
 using MessagePtr = std::shared_ptr<const chatterino::Message>;
 using ContextMap = QMap<QString, QVariant>;
@@ -72,72 +72,20 @@ QString tokenTypeToInfoString(TokenType type);
 class PossibleType
 {
 public:
-    PossibleType(QMetaType::Type t)
-    {
-        this->types_.insert(t);
-    }
+    PossibleType(QMetaType::Type t);
+    PossibleType(std::initializer_list<QMetaType::Type> t);
 
-    PossibleType(std::initializer_list<QMetaType::Type> t)
-        : types_(t)
-    {
-        assert(!this->types_.empty());
-    }
+    QString string() const;
 
-    QString string() const
-    {
-        if (this->types_.size() == 1)
-        {
-            return metaTypeToString(*this->types_.begin());
-        }
-        else
-        {
-            QStringList names;
-            names.reserve(this->types_.size());
-            for (QMetaType::Type t : this->types_)
-            {
-                names.push_back(metaTypeToString(t));
-            }
-            return "(" + names.join(" | ") + ")";
-        }
-    }
-
-    bool operator==(QMetaType::Type t) const
-    {
-        return this->types_.count(t) != 0;
-    }
-
-    bool operator==(const PossibleType &p) const
-    {
-        // Check if there are any common types between the two sets
-        auto i = this->types_.cbegin();
-        auto j = p.types_.cbegin();
-        while (i != this->types_.cend() && j != p.types_.cend())
-        {
-            if (*i == *j)
-                return true;
-            else if (*i < *j)
-                ++i;
-            else
-                ++j;
-        }
-        return false;
-    }
-
-    bool operator!=(QMetaType::Type t) const
-    {
-        return this->types_.count(t) == 0;
-    }
-
-    bool operator!=(const PossibleType &p) const
-    {
-        return !this->operator==(p);
-    }
+    bool operator==(QMetaType::Type t) const;
+    bool operator==(const PossibleType &p) const;
+    bool operator!=(QMetaType::Type t) const;
+    bool operator!=(const PossibleType &p) const;
 
 private:
     std::set<QMetaType::Type> types_;
 };
 
-class Expression;
 class TypeValidator
 {
 public:
@@ -159,121 +107,24 @@ private:
     QString failureMessage_ = "";
 };
 
-class Expression
+inline bool variantIs(const QVariant &a, QMetaType::Type type)
 {
-public:
-    virtual ~Expression() = default;
+    return static_cast<QMetaType::Type>(a.type()) == type;
+}
 
-    virtual QVariant execute(const ContextMap &) const
-    {
-        return false;
-    }
-
-    virtual PossibleType returnType() const
-    {
-        return QMetaType::Bool;
-    }
-
-    virtual bool validateTypes(TypeValidator &validator) const
-    {
-        return true;
-    }
-
-    virtual QString debug() const
-    {
-        return "(false)";
-    }
-
-    virtual QString filterString() const
-    {
-        return "";
-    }
-};
-
-using ExpressionPtr = std::unique_ptr<Expression>;
-
-class ValueExpression : public Expression
+inline bool variantIsNot(const QVariant &a, QMetaType::Type type)
 {
-public:
-    ValueExpression(QVariant value, TokenType type);
-    TokenType type();
+    return static_cast<QMetaType::Type>(a.type()) != type;
+}
 
-    QVariant execute(const ContextMap &context) const override;
-    PossibleType returnType() const override;
-    bool validateTypes(TypeValidator &validator) const override;
-    QString debug() const override;
-    QString filterString() const override;
-
-private:
-    QVariant value_;
-    TokenType type_;
-};
-
-class RegexExpression : public Expression
+inline bool convertVariantTypes(QVariant &a, QVariant &b, int type)
 {
-public:
-    RegexExpression(QString regex, bool caseInsensitive);
+    return a.convert(type) && b.convert(type);
+}
 
-    QVariant execute(const ContextMap &context) const override;
-    PossibleType returnType() const override;
-    bool validateTypes(TypeValidator &validator) const override;
-    QString debug() const override;
-    QString filterString() const override;
-
-private:
-    QString regexString_;
-    bool caseInsensitive_;
-    QRegularExpression regex_;
-};
-
-using ExpressionList = std::vector<std::unique_ptr<Expression>>;
-
-class ListExpression : public Expression
+inline bool variantTypesMatch(QVariant &a, QVariant &b, QMetaType::Type type)
 {
-public:
-    ListExpression(ExpressionList list);
-
-    QVariant execute(const ContextMap &context) const override;
-    PossibleType returnType() const override;
-    bool validateTypes(TypeValidator &validator) const override;
-    QString debug() const override;
-    QString filterString() const override;
-
-private:
-    ExpressionList list_;
-};
-
-class BinaryOperation : public Expression
-{
-public:
-    BinaryOperation(TokenType op, ExpressionPtr left, ExpressionPtr right);
-
-    QVariant execute(const ContextMap &context) const override;
-    PossibleType returnType() const override;
-    bool validateTypes(TypeValidator &validator) const override;
-    QString debug() const override;
-    QString filterString() const override;
-
-private:
-    TokenType op_;
-    ExpressionPtr left_;
-    ExpressionPtr right_;
-};
-
-class UnaryOperation : public Expression
-{
-public:
-    UnaryOperation(TokenType op, ExpressionPtr right);
-
-    QVariant execute(const ContextMap &context) const override;
-    PossibleType returnType() const override;
-    bool validateTypes(TypeValidator &validator) const override;
-    QString debug() const override;
-    QString filterString() const override;
-
-private:
-    TokenType op_;
-    ExpressionPtr right_;
-};
+    return variantIs(a, type) && variantIs(b, type);
+}
 
 }  // namespace filterparser
