+++ src/controllers/filters/parser/expressions/BinaryOperation.cpp
@@ -1,0 +1,390 @@
+#include "BinaryOperation.hpp"
+
+#include <QRegularExpression>
+
+namespace filterparser {
+
+BinaryOperation::BinaryOperation(TokenType op, ExpressionPtr left,
+                                 ExpressionPtr right)
+    : op_(op)
+    , left_(std::move(left))
+    , right_(std::move(right))
+{
+}
+
+QVariant BinaryOperation::execute(const ContextMap &context) const
+{
+    auto left = this->left_->execute(context);
+    auto right = this->right_->execute(context);
+    switch (this->op_)
+    {
+        case PLUS:
+            if (static_cast<QMetaType::Type>(left.type()) ==
+                    QMetaType::QString &&
+                right.canConvert(QMetaType::QString))
+            {
+                return left.toString().append(right.toString());
+            }
+            if (convertVariantTypes(left, right, QMetaType::Int))
+            {
+                return left.toInt() + right.toInt();
+            }
+            return 0;
+        case MINUS:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() - right.toInt();
+            return 0;
+        case MULTIPLY:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() * right.toInt();
+            return 0;
+        case DIVIDE:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() / right.toInt();
+            return 0;
+        case MOD:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() % right.toInt();
+            return 0;
+        case OR:
+            if (convertVariantTypes(left, right, QMetaType::Bool))
+                return left.toBool() || right.toBool();
+            return false;
+        case AND:
+            if (convertVariantTypes(left, right, QMetaType::Bool))
+                return left.toBool() && right.toBool();
+            return false;
+        case EQ:
+            if (variantTypesMatch(left, right, QMetaType::QString))
+            {
+                return left.toString().compare(right.toString(),
+                                               Qt::CaseInsensitive) == 0;
+            }
+            return left == right;
+        case NEQ:
+            if (variantTypesMatch(left, right, QMetaType::QString))
+            {
+                return left.toString().compare(right.toString(),
+                                               Qt::CaseInsensitive) != 0;
+            }
+            return left != right;
+        case LT:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() < right.toInt();
+            return false;
+        case GT:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() > right.toInt();
+            return false;
+        case LTE:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() <= right.toInt();
+            return false;
+        case GTE:
+            if (convertVariantTypes(left, right, QMetaType::Int))
+                return left.toInt() >= right.toInt();
+            return false;
+        case CONTAINS:
+            if (variantIs(left, QMetaType::QStringList) &&
+                right.canConvert(QMetaType::QString))
+            {
+                return left.toStringList().contains(right.toString(),
+                                                    Qt::CaseInsensitive);
+            }
+
+            if (variantIs(left.type(), QMetaType::QVariantMap) &&
+                right.canConvert(QMetaType::QString))
+            {
+                return left.toMap().contains(right.toString());
+            }
+
+            if (variantIs(left.type(), QMetaType::QVariantList))
+            {
+                return left.toList().contains(right);
+            }
+
+            if (left.canConvert(QMetaType::QString) &&
+                right.canConvert(QMetaType::QString))
+            {
+                return left.toString().contains(right.toString(),
+                                                Qt::CaseInsensitive);
+            }
+
+            return false;
+        case STARTS_WITH:
+            if (variantIs(left.type(), QMetaType::QStringList) &&
+                right.canConvert(QMetaType::QString))
+            {
+                auto list = left.toStringList();
+                return !list.isEmpty() &&
+                       list.first().compare(right.toString(),
+                                            Qt::CaseInsensitive);
+            }
+
+            if (variantIs(left.type(), QMetaType::QVariantList))
+            {
+                return left.toList().startsWith(right);
+            }
+
+            if (left.canConvert(QMetaType::QString) &&
+                right.canConvert(QMetaType::QString))
+            {
+                return left.toString().startsWith(right.toString(),
+                                                  Qt::CaseInsensitive);
+            }
+
+            return false;
+
+        case ENDS_WITH:
+            if (variantIs(left.type(), QMetaType::QStringList) &&
+                right.canConvert(QMetaType::QString))
+            {
+                auto list = left.toStringList();
+                return !list.isEmpty() &&
+                       list.last().compare(right.toString(),
+                                           Qt::CaseInsensitive);
+            }
+
+            if (variantIs(left.type(), QMetaType::QVariantList))
+            {
+                return left.toList().endsWith(right);
+            }
+
+            if (left.canConvert(QMetaType::QString) &&
+                right.canConvert(QMetaType::QString))
+            {
+                return left.toString().endsWith(right.toString(),
+                                                Qt::CaseInsensitive);
+            }
+
+            return false;
+        case MATCH: {
+            if (!left.canConvert(QMetaType::QString))
+            {
+                return false;
+            }
+
+            auto matching = left.toString();
+
+            switch (static_cast<QMetaType::Type>(right.type()))
+            {
+                case QMetaType::QRegularExpression: {
+                    return right.toRegularExpression()
+                        .match(matching)
+                        .hasMatch();
+                }
+                case QMetaType::QVariantList: {
+                    auto list = right.toList();
+
+                    // list must be two items
+                    if (list.size() != 2)
+                        return false;
+
+                    // list must be a regular expression and an int
+                    if (variantIsNot(list.at(0),
+                                     QMetaType::QRegularExpression) ||
+                        variantIsNot(list.at(1), QMetaType::Int))
+                        return false;
+
+                    auto match =
+                        list.at(0).toRegularExpression().match(matching);
+
+                    // if matched, return nth capture group. Otherwise, return false
+                    if (match.hasMatch())
+                        return match.captured(list.at(1).toInt());
+                    else
+                        return false;
+                }
+                default:
+                    return false;
+            }
+        }
+        default:
+            return false;
+    }
+}
+
+PossibleType BinaryOperation::returnType() const
+{
+    auto left = this->left_->returnType();
+    auto right = this->right_->returnType();
+    switch (this->op_)
+    {
+        case PLUS:
+            if (left == QMetaType::QString)
+            {
+                return QMetaType::QString;  // String concatenation
+            }
+            return QMetaType::Int;
+        case MINUS:
+        case MULTIPLY:
+        case DIVIDE:
+        case MOD:
+            return QMetaType::Int;
+        case OR:
+        case AND:
+        case EQ:
+        case NEQ:
+        case LT:
+        case GT:
+        case LTE:
+        case GTE:
+        case CONTAINS:
+        case STARTS_WITH:
+        case ENDS_WITH:
+            return QMetaType::Bool;
+        case MATCH: {
+            if (left != QMetaType::QString)
+            {
+                return QMetaType::Bool;
+            }
+
+            if (right == QMetaType::QRegularExpression)
+            {
+                return QMetaType::Bool;
+            }
+            else if (right == QMetaType::QVariantList)
+            {
+                return {QMetaType::QString, QMetaType::Bool};
+            }
+
+            return QMetaType::Bool;
+        }
+        default:
+            return QMetaType::Bool;
+    }
+}
+
+bool BinaryOperation::validateTypes(TypeValidator &validator) const
+{
+    if (!this->left_->validateTypes(validator) ||
+        !this->right_->validateTypes(validator))
+    {
+        return false;
+    }
+
+    auto left = this->left_->returnType();
+    auto right = this->right_->returnType();
+    switch (this->op_)
+    {
+        case PLUS:
+            if (left == QMetaType::QString)
+            {
+                return true;
+            }
+            return validator.must(
+                left == QMetaType::Int && right == QMetaType::Int, this->op_,
+                left, right, this);
+        case MINUS:
+        case MULTIPLY:
+        case DIVIDE:
+        case MOD:
+        case LT:
+        case GT:
+            return validator.must(
+                left == QMetaType::Int && right == QMetaType::Int, this->op_,
+                left, right, this);
+        case OR:
+        case AND:
+            return validator.must(
+                left == QMetaType::Bool && right == QMetaType::Bool, this->op_,
+                left, right, this);
+        case EQ:
+        case NEQ:
+            // todo:
+            // validator.must(left == right || left == QMetaType::QString,
+            //                this->op_, left, right, this);
+            return true;
+        case LTE:
+        case GTE:
+            return validator.must(
+                left == QMetaType::Int && right == QMetaType::Int, this->op_,
+                left, right, this);
+        case CONTAINS:
+        case STARTS_WITH:
+        case ENDS_WITH:
+            return validator.must(left == QMetaType::QVariantList ||
+                                      left == QMetaType::QStringList ||
+                                      left == QMetaType::QString,
+                                  this->op_, left, right, this);
+        case MATCH: {
+            if (left != QMetaType::QString)
+            {
+                validator.fail(
+                    QStringLiteral(
+                        "Can't match on type %1, only string\n\nExpression: %s")
+                        .arg(left.string(), this->filterString()));
+                return false;
+            }
+
+            if (right == QMetaType::QRegularExpression ||
+                right == QMetaType::QVariantList)
+            {
+                return true;
+            }
+
+            return validator.must(false, this->op_, left, right, this);
+        }
+        default:
+            return false;
+    }
+}
+
+QString BinaryOperation::debug() const
+{
+    return QString("(%1 %2 %3)")
+        .arg(this->left_->debug(), tokenTypeToInfoString(this->op_),
+             this->right_->debug());
+}
+
+QString BinaryOperation::filterString() const
+{
+    const auto opText = [&]() -> QString {
+        switch (this->op_)
+        {
+            case AND:
+                return "&&";
+            case OR:
+                return "||";
+            case PLUS:
+                return "+";
+            case MINUS:
+                return "-";
+            case MULTIPLY:
+                return "*";
+            case DIVIDE:
+                return "/";
+            case MOD:
+                return "%";
+            case EQ:
+                return "==";
+            case NEQ:
+                return "!=";
+            case LT:
+                return "<";
+            case GT:
+                return ">";
+            case LTE:
+                return "<=";
+            case GTE:
+                return ">=";
+            case CONTAINS:
+                return "contains";
+            case STARTS_WITH:
+                return "startswith";
+            case ENDS_WITH:
+                return "endswith";
+            case MATCH:
+                return "match";
+            default:
+                return QString();
+        }
+    }();
+
+    return QString("(%1) %2 (%3)")
+        .arg(this->left_->filterString())
+        .arg(opText)
+        .arg(this->right_->filterString());
+}
+
+}  // namespace filterparser
